add favicon alert code
Adds two libraries; a polyfill for the Page Visiblity API and a library for adding alert numbers to the favicon. The favicon alert is only applied when a user's tab is not focused, and disappears when the tab is focused. It does not alert while you're on other pages of the site.

Conflicts:
	public/js/global.js